+++ CHANGELOG.md
@@ -9,6 +9,7 @@ to a section with the version name.
 ## Unreleased Changes
 
 * Hot fix for customer creation on frappe v15.38.0 ([Issue](https://github.com/lavaloon-eg/ksa_compliance/issues/86))
+* Fix showing done only without displaying error message on Perform Compliance Check when there is an issue in company setup.
 
 ## 0.24.1